+++ noxfile.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env -S uv run --script --quiet
 # Copyright (c) 2023 - 2025 Chair for Design Automation, TUM
 # All rights reserved.
 #
@@ -5,34 +6,48 @@
 #
 # Licensed under the MIT License
 
+# /// script
+# dependencies = ["nox"]
+# ///
+
 """Nox sessions."""
 
 from __future__ import annotations
 
 import argparse
+import contextlib
 import os
 import shutil
+import tempfile
 from typing import TYPE_CHECKING
 
 import nox
 
 if TYPE_CHECKING:
-    from collections.abc import Sequence
+    from collections.abc import Generator, Sequence
 
-nox.needs_version = ">=2024.3.2"
+nox.needs_version = ">=2025.10.16"
 nox.options.default_venv_backend = "uv"
 
-nox.options.sessions = ["lint", "tests", "minimums"]
 
-# TODO(denialhaag): Add 3.14 when all dependencies support it  # noqa: FIX002
-#   https://github.com/munich-quantum-toolkit/yaqs/issues/141
-PYTHON_ALL_VERSIONS = ["3.9", "3.10", "3.11", "3.12", "3.13"]
+PYTHON_ALL_VERSIONS = ["3.10", "3.11", "3.12", "3.13", "3.14"]
 
 if os.environ.get("CI", None):
     nox.options.error_on_missing_interpreters = True
 
 
-@nox.session(reuse_venv=True)
+@contextlib.contextmanager
+def preserve_lockfile() -> Generator[None]:
+    """Preserve the lockfile by moving it to a temporary directory."""
+    with tempfile.TemporaryDirectory() as temp_dir_name:
+        shutil.move("uv.lock", f"{temp_dir_name}/uv.lock")
+        try:
+            yield
+        finally:
+            shutil.move(f"{temp_dir_name}/uv.lock", "uv.lock")
+
+
+@nox.session(reuse_venv=True, default=True)
 def lint(session: nox.Session) -> None:
     """Run the linter."""
     if shutil.which("pre-commit") is None:
@@ -68,23 +83,23 @@ def _run_tests(
     )
 
 
-@nox.session(reuse_venv=True, python=PYTHON_ALL_VERSIONS)
+@nox.session(python=PYTHON_ALL_VERSIONS, reuse_venv=True, default=True)
 def tests(session: nox.Session) -> None:
     """Run the test suite."""
     _run_tests(session)
 
 
-@nox.session(reuse_venv=True, venv_backend="uv", python=PYTHON_ALL_VERSIONS)
+@nox.session(python=PYTHON_ALL_VERSIONS, reuse_venv=True, venv_backend="uv", default=True)
 def minimums(session: nox.Session) -> None:
     """Test the minimum versions of dependencies."""
-    _run_tests(
-        session,
-        install_args=["--resolution=lowest-direct"],
-        run_args=["-Wdefault"],
-    )
-    env = {"UV_PROJECT_ENVIRONMENT": session.virtualenv.location}
-    session.run("uv", "tree", "--frozen", env=env)
-    session.run("uv", "lock", "--refresh", env=env)
+    with preserve_lockfile():
+        _run_tests(
+            session,
+            install_args=["--resolution=lowest-direct"],
+            run_args=["-Wdefault"],
+        )
+        env = {"UV_PROJECT_ENVIRONMENT": session.virtualenv.location}
+        session.run("uv", "tree", "--frozen", env=env)
 
 
 @nox.session(reuse_venv=True)
@@ -118,3 +133,7 @@ def docs(session: nox.Session) -> None:
         *shared_args,
         env=env,
     )
+
+
+if __name__ == "__main__":
+    nox.main()